% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{CentralDifference}
\alias{CentralDifference}
\alias{CentralDifference,numeric-method}
\alias{CentralDifference,data.frame-method}
\title{S4 Generic: CentralDifference}
\usage{
\S4method{CentralDifference}{numeric}(x, samplingRate, n, order)

\S4method{CentralDifference}{data.frame}(x, samplingRate, n, order, suffix)
}
\arguments{
\item{x}{A numeric vector or a data frame of numeric vectors.}

\item{samplingRate}{An atomic numeric, the sampling rate of \code{x}, in
hertz. The raw central difference of \code{x} is divided by the sampling
rate so that the central difference can be interepreted as an estimate of
the derivative of \code{x}. Default is \code{1}.}

\item{n}{An atomic numeric, the number of samples offset from the current
sample used in the central difference. Default is \code{1}.}

\item{order}{An atomic numeric, the number of times the central-difference
operator is applied to \code{x}. Default is \code{1}.}

\item{suffix}{A character string. If \code{x} is a data frame, the names of the
returned data frame are equal to the names of \code{x}, each suffixed by
\code{suffix}. Default is \code{""}.}
}
\value{
The central difference of \code{x} if \code{x} is numeric, or of
  each column of \code{x} if \code{x} is a data frame.
}
\description{
Given a numeric series \code{x}, compute its \code{n}-offset central
difference \code{y}, such that \code{y[i] = x[i+n] - x[i-n]}.
}
