% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{FindLandmarks}
\alias{FindLandmarks}
\alias{FindLandmarks.data.frame}
\alias{FindLandmarks.list}
\title{Find Gestural Landmarks}
\usage{
FindLandmarks(x, ...)

\method{FindLandmarks}{data.frame}(x, channels = NULL, matches = NULL,
  from = -Inf, to = Inf, onsetNear = NULL, targetNear = NULL,
  theta = 0.2, alpha = 0.1, ...)

\method{FindLandmarks}{list}(x, channels = NULL, matches = NULL,
  from = -Inf, to = Inf, onsetNear = NULL, targetNear = NULL,
  theta = 0.2, alpha = 0.1, ...)
}
\arguments{
\item{x}{A data.frame (or a list of data.frames) that contain(s) kinematic
data (e.g., position, velocity, speed, or acceleration data). \code{x}
must have one column named \code{Time} and at least one other column.}

\item{...}{Placeholder for future methods.}

\item{channels}{A character vector, each element of which should match
exactly the name of some column in \code{x} other than \code{Time}.
Default matches all columns of \code{x} other than \code{Time}.}

\item{matches}{A pattern used to match columns in \code{x}. A non-null
value for this argument will supercede a non-null value for the
\code{channels} argument. Default matches all columns of \code{x} other
than \code{Time}.}

\item{from, to}{Atomic numerics. Local maxima will only be found within the
\code{[from,to]} interval. Default values will include all available
data in the channel.}

\item{onsetNear, targetNear}{Atomic numerics.}

\item{theta}{A proportion within the range \code{[0,1)}. Given consecutive
local maximum \code{max} and local minimum \code{min}, the gestural
landmark \code{mark} is the point in time between the local maximum
and minimum such that:
\code{value(mark) = value(min) + (theta * (value(max) - value(min)))}.
Default is \code{0.2}, which is the "industry standard".}

\item{alpha}{A proportion within the range \code{[0,1)}. Local maxima in a
given channel will only be found if their value exceeds \code{100*alpha}
percent of the data range in that channel.}
}
\value{
A data table with the following variables:
  \itemize{
    \item \code{Channel}: the name of the channel in which the local
      maximum occurs.
    \item \code{Landmark}: the name of the landmark: \code{GestureOnset},
      \code{TargetOnset}, \code{TargetOffset}, \code{GestureOffset}.
    \item \code{Time}: the time at which the local occurs.
    \item \code{Value}: the amplitude value of the local minimum.
  }
}
\description{
Find the gestural onset, target onset, target offset, and gestural offset
within a speed time-series.
}
\details{
Gestural landmarks may be found only within a time series that denotes the
speed of a sensor (or of a derived quantity, such as lip aperture) in one
or more dimensions. Within a speed time-series, a gesture is comprises three
local minima (\code{m1}, \code{m3}, \code{m5}), with a local maximum
(\code{m2}, \code{m4}) between each local minima: i.e., these local minima
and maxima are ordered such that \code{m1 < m2 < m3 < m4 < m5}. The first
half of the gesture, during which the constriction is formed, corresponds to
the interval \code{[m1, m3)}, where the speed values trace a concave downward
trajectory. The second half of the gesture, during which the constriction
is released, corresponds to the interval \code{[m3, m5]}, where again the
speed values trace a concave downward trajectory.

A gestural landmark corresponds to the point in time between two consecutive
local extrema, where the speed time-series is equal to some proportion
(\code{theta}) of the range between the two local extrema. For example, if
\code{theta = 0.2}, then the gestural onset is the point in time where the
speed is equal to 20 percent of the range between the speed at \code{m1},
and the speed at \code{m2}: \code{speed(m1) + (0.2 * (speed(m2) - speed(m1)))}.

In order to find landmarks, the user must provide a fixed point in time that
is near to the local minima of the onset of the gesture (i.e., via
the \code{onsetNear} argument) or the target (i.e., via the \code{targetNear}
argument).
}
