% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{FindMaxima}
\alias{FindMaxima}
\alias{FindMaxima.data.frame}
\alias{FindMaxima.list}
\title{Find Local Maxima}
\usage{
FindMaxima(x, ...)

\method{FindMaxima}{data.frame}(x, channels = NULL, matches = NULL,
  from = -Inf, to = Inf, alpha = 0.1, ...)

\method{FindMaxima}{list}(x, channels = NULL, matches = NULL, from = -Inf,
  to = Inf, alpha = 0.1, ...)
}
\arguments{
\item{x}{A data.frame (or a list of data.frames) that contain(s) kinematic
data (e.g., position, velocity, speed, or acceleration data). \code{x}
must have one column named \code{Time} and at least one other column.}

\item{...}{Placeholder for future methods.}

\item{channels}{A character vector, each element of which should match
exactly the name of some column in \code{x} other than \code{Time}.
Default matches all columns of \code{x} other than \code{Time}.}

\item{matches}{A pattern used to match columns in \code{x}. A non-null
value for this argument will supercede a non-null value for the
\code{channels} argument. Default matches all columns of \code{x} other
than \code{Time}.}

\item{from, to}{Atomic numerics. Local maxima will only be found within the
\code{[from,to]} interval. Default values will include all available
data in the channel.}

\item{alpha}{A proportion within the range \code{[0,1)}. Local maxima in a
given channel will only be found if their value exceeds \code{100*alpha}
percent of the data range in that channel.}
}
\value{
A data table with the following variables:
  \itemize{
    \item \code{Channel}: the name of the channel in which the local
      maximum occurs.
    \item \code{Time}: the time at which the local occurs.
    \item \code{Value}: the amplitude value of the local maximum.
  }
}
\description{
Find the local maxima within one or more channels of kinematic data.
}
