% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{FormatChannels}
\alias{FormatChannels}
\title{Format Channel Names}
\usage{
FormatChannels(sensors = c(""), n = 16)
}
\arguments{
\item{sensors}{A character vector, names for sensors that were attached to
the participant and connected to sockets in the Carstens Sensin box. The
position of a name in this vector is interpreted as the number of the
socket to which the sensor was connected. Default is \code{c("")}.}

\item{n}{An atomic numeric, the number of sensors/sockets for which the
Carstens Sweepsaver software recorded data. (This number is recorded in
the \code{NumberOfChannels} field in the header of a \code{.pos} file.)
Default is \code{16}.}
}
\value{
A character vector, the result of first formatting \code{sensors}
  (by calling \code{FormatSensors}) and then suffixing each formatted
  sensor name with \code{x}, \code{y}, \code{z}, \code{phi}, \code{theta},
  \code{rms}, and \code{extra}.
}
\description{
Expand a character vector whose elements denote sensor names into a character
vector whose elements denote names for the channels of data that were
recorded for those sensors.
}
\details{
For each sensor, seven channels of data are recorded:
\itemize{
  \item \code{x}: position of the sensor along the antero-posterior axis
                  (i.e., orthogonal to the coronal plane)
  \item \code{y}: position of the sensor along the lateral axis
                  (i.e., orthogonal to the sagittal plane)
  \item \code{z}: position of the sensor along the vertical axis
                  (i.e., orthogonal to the transverse plane)
  \item \code{phi}: rotation angle between the \code{x}-axis and the
                    projection of the sensor axis onto the transverse plane
                    (i.e., the azimuthal angle)
  \item \code{theta}: rotation angle between the \code{z}-axis and the
                      sensor axis (i.e., the polar angle)
  \item \code{rms}: the root-mean-squared error between the measured
                    amplitudes from the articulograph's transmitter coils
                    and the expected amplitudes given the positional
                    and rotational coordinates of the sensor
  \item \code{extra}: an empty channel that serves as a placeholder for
                      future development
}
The \code{x}, \code{y}, and \code{z} axes define a three-dimensional
right-handed Cartesian coordinate system. The \code{phi} and \code{theta}
angles define the sensor axis in a spherical coordinate system.

Because seven channels of data are recorded for each sensor, the returned
character vector has \code{7n} elements.

Note that the MATLAB toolbox SMASH names the positional axes differently:
the antero-posterior axis is named \code{z}; the lateral axis, \code{x}; and
the vertical axis, \code{y}.
}
\examples{
FormatChannels(n = 8)
FormatChannels(sensors = c("HL", "HR", "TB", "TD", "TT"), n = 16)
FormatChannels(sensors = c("", "", "TB", "TD", "TT"), n = 16)

}
\seealso{
\code{\link{FormatSensors}}
}
