% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensors.R
\name{FormatSensors}
\alias{FormatSensors}
\title{Format Sensor Names}
\usage{
FormatSensors(sensors = c(""), n = 16)
}
\arguments{
\item{sensors}{A character vector, names for sensors that were attached to
the participant and connected to sockets in the Carstens Sensin box. The
position of a name in this vector is interpreted as the number of the
socket to which the sensor was connected. Default is \code{c("")}.}

\item{n}{An atomic numeric, the number of sensors/sockets for which the
Carstens Sweepsaver software recorded data. (This number is recorded in
the \code{NumberOfChannels} field in the header of a \code{.pos} file.)
Default is \code{16}.}
}
\value{
A character vector, derived by padding \code{sensors} with empty
  strings (\code{""}) to length \code{n}, and then replacing the empty
  strings with \code{"CH##"}, where \code{##} denotes the position of the
  string.
}
\description{
Format a character vector whose elements denote sensor names.
}
\examples{
FormatSensors(n = 8)
FormatSensors(sensors = c("HL", "HR", "TB", "TD", "TT"), n = 16)
FormatSensors(sensors = c("", "", "TB", "TD", "TT"), n = 16)

}
