% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{ForwardDifference}
\alias{ForwardDifference}
\alias{ForwardDifference.numeric}
\alias{ForwardDifference.data.frame}
\alias{ForwardDifference.list}
\alias{CentralDifference.list}
\title{Forward Difference}
\usage{
ForwardDifference(x, ...)

\method{ForwardDifference}{numeric}(x, n = 1, order = 1, samplingRate = n,
  ...)

\method{ForwardDifference}{data.frame}(x, n = 1, order = 1, ...)

\method{ForwardDifference}{list}(x, n = 1, order = 1, ...)

\method{CentralDifference}{list}(x, n = 1, order = 1, ...)
}
\arguments{
\item{x}{A numeric vector or a data frame of numeric vectors.}

\item{...}{Placeholder for future methods.}

\item{n}{An atomic numeric, the number of samples forward from the current
sample used in the forward difference. Default is \code{1}.}

\item{order}{An atomic numeric, the number of times the forward-difference
operator is applied to \code{x}. Default is \code{1}.}

\item{samplingRate}{An atomic numeric, the sampling rate of \code{x}, in
hertz. The raw forward difference of \code{x} is divided by the sampling
rate so that the forward difference can be interepreted as an estimate of
the derivative of \code{x}. Default is \code{n}, which yields the raw
forward difference.}
}
\value{
The forward difference of \code{x} if \code{x} is numeric, or of
  each column of \code{x} if \code{x} is a data frame.
}
\description{
Given a numeric series \code{x}, compute its \code{n}-offset forward
difference \code{y}, such that \code{y[i] = x[i+n] - x[i]}.
}
