% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greek-clusters.R
\name{GreekClusters}
\alias{GreekClusters}
\title{Greek Clusters}
\usage{
GreekClusters(includePaths = TRUE)
}
\arguments{
\item{includePaths}{If \code{TRUE}, then the paths to the \code{.pos} and
  \code{.txt} files in the installed package are found, and included as
  variables named \code{POS} and \code{TXT}, respectively.

  If \code{FALSE}, then the paths to these files for the sweep are not
  included.}
}
\value{
If \code{includePaths = TRUE}, a data table with 4 rows and
  9 variables:
  \itemize{
    \item \code{Sweep}: the sweep number in which the cluster was produced
      and recorded.
    \item \code{Word}, \code{Cluster}, \code{Vowel}, \code{Anchor}:
      the WorldBet transcriptions of the target word, the word-initial
      fricative-stop cluster, the post-cluster vowel, and the post-vocalic
      singleton consonant, respectively.
    \item \code{Onset}, \code{Offset}:
      the approximate times of the onset and offset of the production
      of the word in the sweep. (Note: These times do not consistently
      correspond to any well-defined acoustic or kinematic events.)
    \item \code{POS}, \code{TXT}:
      the paths to the \code{.pos} and \code{.txt} files in the installed
      package for the sweeps.
  }

  If \code{includePaths = FALSE}, then returned data table omits the
  \code{POS} and \code{TXT} variables
}
\description{
Metadata for productions of Greek /sk/ and /sp/ clusters.
}
