% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ListSweeps}
\alias{ListSweeps}
\title{List Sweeps}
\usage{
ListSweeps(path = ".", pattern = "*", fullNames = TRUE, simplify = TRUE,
  recursive = FALSE, dropMissing = TRUE)
}
\arguments{
\item{path}{A character vector of paths to directories that contain sweeps
(pairs of \code{.pos} and \code{.txt} files that share the same basename).
Tilde expansion (see \code{\link{path.expand}}) is performed.
Default value corresponds to the present working directory.}

\item{pattern}{An optional vector of \code{\link[=regex]{regular expressions}}.
The length of this vector should be equal to the length of the \code{path}
argument, or equal to \code{1}, in which case the same regular expression
is used to match sweeps in each directory listed in \code{path}.
Only sweeps whose basenames match the regular expression(s) will be
returned. Default value matches all available sweeps. Patterns are always
case-sensitive.}

\item{fullNames}{If \code{TRUE}, then the full paths for the \code{.pos}
and \code{.txt} files are returned. If \code{FALSE}, then the basenames
for these files are returned. Default is \code{TRUE}.}

\item{simplify}{If \code{TRUE}, then the table of sweeps found for each element
in \code{path} are row-binded together into a single data table.
If \code{FALSE}, then the tables of sweeps found for the elements in \code{path}
are returned as elements of a list.}

\item{recursive}{If \code{TRUE}, then the search recurses into subdirectories of
elements of \code{param}. If \code{FALSE}, then the search only looks
immediately in the directories listed in \code{path}.}

\item{dropMissing}{If \code{TRUE}, then sweeps that are missing either a
\code{.pos} or a \code{.txt} file are not listed. If \code{FALSE}, then
sweeps with a missing file are listed.}
}
\value{
If \code{simplify = TRUE}, a data table with the following variables:
  \itemize{
    \item \code{Sweep}: the basenames of the sweeps.
    \item \code{POS}: the \code{.pos} files for the sweeps.
    \item \code{TXT}: the \code{.txt} files for the sweeps.
  }

  If \code{simplify = FALSE}, a list with as many elements as \code{path}.
  Each element is a data table as described immediately above.
}
\description{
List the \code{.pos} and \code{.txt} files for sweeps. Vectorized over
\code{path} and \code{pattern} arguments.
}
