% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{ReadBatch}
\alias{ReadBatch}
\title{Read the Positional Data for a Batch of Sweeps}
\usage{
ReadBatch(path, sensors = c(""))
}
\arguments{
\item{path}{A character string, the path to the directory containing the
\code{.txt} and \code{.pos} files.}

\item{sensors}{A character vector, names for sensors that were attached to
the participant and connected to sockets in the Carstens Sensin box. The
position of a name in this vector is interpreted as the number of the
socket to which the sensor was connected. Default is \code{c("")}.}
}
\value{
A data tibble with the same variables as those returned by
  \code{ReadPos}.
}
\description{
Read head-corrected positional data from \code{.txt} files and metadata
from the headers of \code{.pos} files.
}
\seealso{
\code{\link{ReadSweep}}
}

