% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{ReadSweep}
\alias{ReadSweep}
\title{Read the Positional Data Recorded During a Sweep}
\usage{
ReadSweep(sweep, path, sensors = c(""))
}
\arguments{
\item{sweep}{A character string, the basename of associated \code{.txt} and
\code{.pos} files that are generated from post-processing a sweep. That is,
there should be files \code{<sweep>.txt} and \code{<sweep>.pos} located in
\code{path}.}

\item{path}{A character string, the path to the directory containing the
\code{.txt} and \code{.pos} files.}

\item{sensors}{A character vector, names for sensors that were attached to
the participant and connected to sockets in the Carstens Sensin box. The
position of a name in this vector is interpreted as the number of the
socket to which the sensor was connected. Default is \code{c("")}.}
}
\value{
A data tibble, with the following variables:
\itemize{
  \item all variables returned by \code{ReadSweepMetadata}
  \item \code{TimeData}: a list of data tibbles, the times at which position
        data samples were recorded
  \item \code{PositionData}: a list of data tibbles, the position data
        returned by \code{ReadSweepData} for the active sensors that were
        recorded during the sweep and then post-processed with Calcpos and
        Normpos
}
}
\description{
Read the head-corrected positional data from a \code{.txt} file and its
metadata from the header of a \code{.pos} file.
}
\seealso{
\code{\link{ReadSweepMetadata}}, \code{\link{ReadSweepData}}
}

