% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadSweepData}
\alias{ReadSweepData}
\title{Read the Contents of the .txt File for a Sweep}
\usage{
ReadSweepData(txt, sensors = character(), n = 7, dropExtra = TRUE)
}
\arguments{
\item{txt}{A character vector of paths to \code{.txt} files.}

\item{sensors}{A character vector or list of character vectors, the
names for sensors that were attached to the participant and connected to
sockets in the Carstens Sensin box. The position of a name in this vector
is interpreted as the number of the socket to which the sensor was connected.
Default is \code{character()}. See \code{\link{FormatSensors}}.}

\item{n}{A numeric, the number of data channels that are recorded
per sensor. Default is \code{7}, which is consistent with Carstens data
format \code{AG50xDATA_V003}, under which each sensor comprises data
channels: \code{x}, \code{y}, \code{z}, \code{phi}, \code{theta},
\code{rms}, \code{extra}.
See \code{\link{FormatChannels}}.}

\item{dropExtra}{If \code{TRUE}, then the \code{extra} channel for each sensor
is dropped from the data, since this channel is only a placeholder for
future development by Carstens and contains no data.
If \code{FALSE}, then the \code{extra} channels are kept.}
}
\value{
If \code{txt} is a single character string, then a data table,
  the names of whose variables are determined by expanding the \code{sensors}
  argument through \code{\link{FormatChannels}}, and optionally dropping
  the channels suffixed by \code{extra}.

  If \code{txt} is a character vector with more than one element, then a list
  whose elements are data tables structured as described above.
}
\description{
Read the contents of a \code{.txt} file, which contains the positional data
recorded during a sweep.
}
\details{
\code{ReadSweepData} is vectorized over both \code{txt} and \code{sensors}.
If \code{txt} is a vector of more than one path to a \code{.txt} file, and
if \code{sensors} is a character vector, then the elements of \code{sensors}
are recycled as the sensor-names for each \code{.txt} file. To specify
different sensor names for each \code{.txt} file, \code{sensors} should be
a list with the same length as \code{txt} and whose elements are character
vectors.
}
