% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.R
\name{ReadSweepMetadata}
\alias{ReadSweepMetadata}
\title{Read the Header of the .pos File for a Sweep}
\usage{
ReadSweepMetadata(file, sensors = c(""))
}
\arguments{
\item{file}{A character string, the path to a \code{.pos} file.}

\item{sensors}{A character vector, names for sensors that were attached to
the participant and connected to sockets in the Carstens Sensin box. The
position of a name in this vector is interpreted as the number of the
socket to which the sensor was connected. Default is \code{c("")}.}
}
\value{
A data tibble with the following variables:
\itemize{
  \item \code{Sweep}
  \item \code{Format}
  \item \code{SamplingRate}
  \item \code{SweepsaverBuild}
  \item \code{SweepsaverVersion}
  \item \code{SweepsaverRevision}
  \item \code{SweepsaverTimestamp}
  \item \code{CalcposBuild}
  \item \code{CalcposVersion}
  \item \code{CalcposRevision}
  \item \code{CalcposTimestamp}
  \item \code{NormposBuild}
  \item \code{NormposVersion}
  \item \code{NormposRevision}
  \item \code{NormposTimestamp}
  \item \code{TaxonomicDistanceMean}
  \item \code{TaxonomicDistanceStdDev}
  \item \code{PostFilters}
}
}
\description{
Read the header of a \code{.pos} file, which contains the metadata for the
positional data recorded during a sweep.
}

