% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\title{Objects Reexported from Other Packages}
\description{
These objects are available through other packages.
Follow the links below to see their documentation.
}
\section{magrittr}{

  \itemize{
    \item \code{\link[magrittr]{\%>\%}}
  }
}

