% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{TimeSlice}
\alias{TimeSlice}
\alias{TimeSlice.data.frame}
\alias{TimeSlice.list}
\title{Time Slice}
\usage{
TimeSlice(x, from, to, ...)

\method{TimeSlice}{data.frame}(x, from, to, ...)

\method{TimeSlice}{list}(x, from, to, ...)
}
\arguments{
\item{x}{A data.frame or a list of data.frames that contain kinematic data,
structured such that there is a column named \code{Time} and other columns
that record the channels of Carstens sensors.}

\item{from}{A numeric, the time from which to slice.}

\item{to}{A numeric, the time to which to slice.}

\item{...}{Placeholder for future methods.}
}
\value{
An object that has the same class as \code{x}. If \code{x} is a
  data.frame, then the returned data.frame is like \code{x}, but filtered
  to just those rows whose value of \code{Time} falls within the
  \code{[from, to]} interval. If \code{x} is a list, then a list of
  time-sliced data.frames is returned.
}
\description{
Slice a contiguous temporal interval of kinematic data.
}
