% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference.R
\name{CentralDifference}
\alias{CentralDifference}
\alias{CentralDifference.numeric}
\alias{CentralDifference.data.frame}
\title{Central Difference}
\usage{
CentralDifference(x, ...)

\method{CentralDifference}{numeric}(x, n = 1, order = 1, samplingRate = 2
  * n, ...)

\method{CentralDifference}{data.frame}(x, n = 1, order = 1, ...)
}
\arguments{
\item{x}{A numeric vector or a data frame of numeric vectors.}

\item{...}{Placeholder for future methods.}

\item{n}{An atomic numeric, the number of samples offset from the current
sample used in the central difference. Default is \code{1}.}

\item{order}{An atomic numeric, the number of times the central-difference
operator is applied to \code{x}. Default is \code{1}.}

\item{samplingRate}{An atomic numeric, the sampling rate of \code{x}, in
hertz. The raw central difference of \code{x} is divided by the sampling
rate so that the central difference can be interepreted as an estimate of
the derivative of \code{x}. Default is \code{2*n}, which yields the raw
central difference.}
}
\value{
The central difference of \code{x} if \code{x} is numeric, or of
  each column of \code{x} if \code{x} is a data frame.
}
\description{
Given a numeric series \code{x}, compute its \code{n}-offset central
difference \code{y}, such that \code{y[i] = x[i+n] - x[i-n]}.
}
