% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadSweep}
\alias{ReadSweep}
\title{Read the Sensor Data Recorded During a Sweep}
\usage{
ReadSweep(pos = "", txt = "", sweep = NULL, sensors = character(),
  dropExtra = TRUE, simplify = TRUE)
}
\arguments{
\item{pos}{A character vector of paths to \code{.pos} files.}

\item{txt}{A character vector of paths to \code{.txt} files.}

\item{sweep}{A character vector, the path to and basenames (without extensions)
of sweeps. The names of the \code{.pos} and \code{.txt} files for the
sweeps will be constructed by appending the appropriate file extensions.
Default is \code{NULL}. If not \code{NULL}, then \code{-.pos} and
\code{-.txt} completions of \code{sweep} override any provided values of
the \code{pos} and \code{txt} arguments.}

\item{sensors}{A character vector, names for sensors that were attached to
the participant and connected to sockets in the Carstens Sensin box. The
position of a name in this vector is interpreted as the number of the
socket to which the sensor was connected. Default is \code{character()}.
See \code{\link{FormatSensors}} and \code{\link{FormatChannels}}.}

\item{dropExtra}{If \code{TRUE}, then the \code{extra} channel for each sensor
is dropped from the data, since this channel is only a placeholder for
future development by Carstens and contains no data.
If \code{FALSE}, then the \code{extra} channels are kept.}

\item{simplify}{If \code{TRUE}, then the data tables for the sweeps
are row-binded together into a single data table.
If \code{FALSE}, then the data tables for the sweeps are returned as
elements of a list.}
}
\value{
A list or nested data frame with the following variables:

\itemize{
  \item \code{Sweep}: A character, the name of the sweep, which identifies
    files with \code{.pos} and \code{.txt} extensions.
  \item \code{Format}: A character, the data format defined by Carstens.
  \item \code{SamplingRate}: A numeric, the number of samples recorded
    per second by the AG501.
  \item \code{SweepsaverBuild}: A character, the version and revision number
    of the Sweepsaver application, which records the amplitude values
    detected by the sensor coils in the AG501.
  \item \code{SweepsaverVersion}: A numeric, the version number of the
    Sweepsaver application.
  \item \code{SweepsaverRevision}: A numeric, the revision number of the
    Sweepsaver application.
  \item \code{SweepsaverTimestamp}: A datetime, the date and time when the
    sweep was recorded by the Sweepsaver application, in
    \code{YYYY-MM-DD HH:MM:SS} format.
  \item \code{CalcposBuild}: A character, the version and revision number
    of the Calcpos application, which transforms the amplitude values
    recorded by Sweepsaver into raw position data that has not been
    corrected for head movement.
  \item \code{CalcposVersion}: A numeric, the version number of the Calcpos
    application.
  \item \code{CalcposRevision}: A numeric, the revision number of the Calcpos
    application.
  \item \code{CalcposTimestamp}: A datetime, the date and time when the
    raw position data was computed by Calcpos.
  \item \code{NormposBuild}: A character, the version and revision number
    of the Normpos application, which transforms the raw position data by
    removing head movement. If the Carstens biteplane was used to compute
    a rotation matrix, then Normpos also rotates and translates the data
    so that the participant's biteplane lies within the x-y plane and so
    that the participant's upper incisors lie at the origin of the coordinate
    system.
  \item \code{NormposVersion}: A numeric, the version number of the Normpos
    application.
  \item \code{NormposRevision}: A numeric, the revision number of the Normpos
    application.
  \item \code{NormposTimestamp}: A datetime, the date and time when the
    head-correction and biteplane-rotation were applied by Normpos.
  \item \code{TaxonomicDistanceMean}: A numeric, the mean of the taxonomic
    distances, averaging across samples. The taxonomic distance of each
    sample is, roughly, the difference between the measured position and the
    expected position relative to a reference object.
  \item \code{TaxonomicDistanceStdDev}: A numeric, the standard deviation of
    taxonomic distances for the sweep.
  \item \code{PostFilters}: A data frame that records the low-pass Kaiser
    filters that were applied to the data channels of the active sensors,
    during postprocessing by the Calcpos and Normpos applications. Includes
    the following columns:
    \itemize{
      \item \code{Socket}: A numeric, the number of the socket in the Sensin
        box to which the physical sensor was connected.
      \item \code{Sensor}: A character, the name of the sensor whose data
        channels were low-pass filtered. See: \code{\link{FormatChannels}}.
      \item \code{Name}: A character, the name of the filter applied to the
        sensor's data channels. See: \code{\link{CarstensFilters}}.
      \item \code{Application}: A character, the application that applied
        the low-pass filter to the data.
    }
}

\itemize{
  \item \code{SensorData}: A list-col, whose only entry is the data frame
    of the position and rotation data for the active sensors, returned by
    \code{\link{ReadSweepData}}.
}
}
\description{
Read the head-corrected (and biteplane-rotated) position and rotation data from a
\code{.txt} file and its metadata from the header of a \code{.pos} file.
}
\details{
\code{ReadSweep} is vectorized over \code{pos}, \code{txt}, and \code{sensors}.
\code{pos} and \code{txt} should be vectors of the same length ordered such
that the positions on these vectors correspond to the \code{.pos} and \code{.txt}
files, respectively, of the same sweep. Alternatively, multiple sweeps can
be specified by passing a character vector to the \code{sweep} argument.
If multiple sweeps are specified, and if \code{sensors} is a character vector,
then the elements of \code{sensors} are recycled as the sensor-names for each
sweep. To specify different sensor names for each sweep, \code{sensors} should be
a list with the same length as the number of sweeps and whose elements are
character vectors.
}
