% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ReadSweepMetadata}
\alias{ReadSweepMetadata}
\title{Read the Header of the .pos File for a Sweep}
\usage{
ReadSweepMetadata(pos, sensors = character())
}
\arguments{
\item{pos}{A character vector of paths to \code{.pos} files.}

\item{sensors}{A character vector or list of character vectors, the
names for sensors that were attached to the participant and connected to
sockets in the Carstens Sensin box. The position of a name in this vector
is interpreted as the number of the socket to which the sensor was connected.
Default is \code{character()}. See \code{\link{FormatSensors}}.}
}
\value{
If \code{pos} is a single character string, then a single-row data
  table with the following variables:
  @template sweep-metadata

  If \code{pos} is a character vector with more than one element, then a list
  whose elements are data tables structured as described above.
}
\description{
Read the header of a \code{.pos} file, which contains the metadata for the
positional data recorded during a sweep.
}
\details{
\code{ReadSweepMetadata} is vectorized over both \code{pos} and \code{sensors}.
If \code{pos} is a vector of more than one path to a \code{.pos} file, and
if \code{sensors} is a character vector, then the elements of \code{sensors}
are recycled as the sensor-names for each \code{.pos} file. To specify
different sensor names for each \code{.pos} file, \code{sensors} should be
a list with the same length as \code{pos} and whose elements are character
vectors.
}
